+++ src/gui/window.rs
@@ -1,4 +1,7 @@
-use crate::prelude::*;
+use gio::SimpleAction;
+use presage::store::Thread;
+
+use crate::{backend::Channel, prelude::*};
 
 glib::wrapper! {
     /// The main application window.
@@ -24,9 +27,30 @@ impl Window {
             );
         }
         app.set_accels_for_action("channel-list.toggle-search", &["<Control>f"]);
-        Object::builder::<Self>()
+
+        let window = Object::builder::<Self>()
             .property("application", app)
-            .build()
+            .build();
+
+        let action_notification_clicked = SimpleAction::new(
+            "notification-clicked",
+            Some(&<(u8, Vec<u8>)>::static_variant_type()),
+        );
+        action_notification_clicked.connect_activate(clone!(
+            #[weak]
+            window,
+            move |_, variant| {
+                let variant: (u8, Vec<u8>) = variant
+                    .expect("Notification-clicked action to have attached data")
+                    .get()
+                    .expect("Variant of notification is of unexpected type");
+                let thread = Channel::from_notification_variant(variant);
+                window.activate_channel(thread);
+            }
+        ));
+        app.add_action(&action_notification_clicked);
+
+        window
     }
 
     // Closes the window, even if hide-on-close is set.
@@ -67,6 +91,13 @@ impl Window {
         if is_maximized {
             self.maximize();
         }
+    }
+
+    fn activate_channel(&self, thread: Thread) {
+        let channel = self
+            .property::<Manager>("manager")
+            .channel_from_thread(thread);
+        self.imp().channel_list.set_active_channel(channel);
     }
 
     pub(crate) fn settings(&self) -> gio::Settings {